UI fixes and object oriented programming shell
I addressed some of the reactive input problems we had with the UI. Made
the number of input modules variable, and hooked up the file inputs.
I also began the shell for the OO version of the starry map logic,
introducing the objects StarryModule, StarryModuleSet, and StarryMap
(with StarryMatrix to come).